Refactor bundled and example file generators
