+++ package.json
@@ -8,6 +8,7 @@
     "preview": "REACT_APP_VERSION=$npm_package_version aio app run",
     "grammar": "nearleyc web-src/src/helpers/expressions.ne -o web-src/src/helpers/Parser.generated.js",
     "prompts": "node scripts/prompt-generator.js",
+    "examples": "node scripts/zip-examples.js",
     "build": "REACT_APP_VERSION=$npm_package_version aio app build",
     "deploy:app": "npm run grammar && REACT_APP_VERSION=$npm_package_version aio app deploy --no-actions",
     "deploy:actions": "aio app deploy --no-web-assets",
